+++ nix/deps.nix
@@ -5,13 +5,12 @@ in
 
 with pkgs; [
   inkscape # For SVG -> PDF
-  graphviz_2_32 # For DOT -> PDF
+  graphviz # For DOT -> PDF
   mscgen # For MSC -> PDF
   imagemagick # For image manipulation (mostly PDF -> PNG)
 
   # Markdown -> PDF:
   pandoc
-  haskellPackages.pandoc-citeproc
   haskellPackages.pandoc-crossref
 
   # TeX: